+++ riberry/com/base.py
@@ -1,5 +1,6 @@
 from enum import IntEnum
 import os.path as osp
+import platform
 
 
 class PacketType(IntEnum):
@@ -27,10 +28,11 @@ class ComBase:
     @staticmethod
     def identify_device():
         try:
-            with open("/proc/cpuinfo") as f:
-                cpuinfo = f.read()
-            if "Raspberry Pi" in cpuinfo:
-                return "Raspberry Pi"
+            if osp.exists("/proc/cpuinfo"):
+                with open("/proc/cpuinfo") as f:
+                    cpuinfo = f.read()
+                if "Raspberry Pi" in cpuinfo:
+                    return "Raspberry Pi"
             if osp.exists("/proc/device-tree/model"):
                 with open("/proc/device-tree/model") as f:
                     model = f.read().strip().replace("\x00", "")
@@ -40,6 +42,10 @@ class ComBase:
                     return model
             if osp.exists('/usr/local/m5stack/block-mount.sh'):
                 return 'm5stack-LLM'
+            if platform.system() == 'Linux':
+                return 'Linux'
+            elif platform.system()== 'Darwin':
+                return 'Darwin'
             return "Unknown Device"
         except FileNotFoundError:
             return "Unknown Device"
